requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
mysql-connector-python